# perf: pkg/util/mapz Enhancement (sub task of pkg/util Enhancement #833)