## Pre-Checklist

Note: please complete **_ALL_** items in the following checklist.

- [x] I have read through the [CONTRIBUTING.md](https://github.com/devstream-io/devstream/blob/main/CONTRIBUTING.md) documentation.
- [x] My code has the necessary comments and documentation (if needed).
- [x] I have added relevant tests

## Description
After investigation and testing, there is a place that could be improved for better performance.

## Related Issues
see #833 for more details

## New Behavior (screenshots if needed)

### before improvement:
```sh
goos: linux
goarch: amd64
pkg: github.com/devstream-io/devstream/pkg/util/mapz
cpu: Intel(R) Xeon(R) CPU E5-1650 0 @ 3.20GHz
BenchmarkFillMapWithStrAndError-12         92894             13964 ns/op           10170 B/op          9 allocs/op
PASS
ok      github.com/devstream-io/devstream/pkg/util/mapz 2.378s
```
### after improvement:
```sh
goos: linux
goarch: amd64
pkg: github.com/devstream-io/devstream/pkg/util/mapz
cpu: Intel(R) Xeon(R) CPU E5-1650 0 @ 3.20GHz
BenchmarkFillMapWithStrAndError-12        157182              6585 ns/op            5401 B/op          3 allocs/op
PASS
ok      github.com/devstream-io/devstream/pkg/util/mapz 1.130s
```